MedDiag-adrian/requirements.txt
streamlit==1.49.1
streamlit-option-menu==0.4.0
scikit-learn==1.7.2

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1